Changes the mapper's map initialization to happen when the component is ready instead of created
This is necessary when using the vue router. if you navigate way from and back to the map component
the map dom element won't be created at the point the component's 'created' lifecycle hook
is fired. Changing it to the 'ready' lifecycle hook asures the dom element does indeed exist.