Removes placemapper hooks from gallery. Adds default route.
I'm leaving placempapper in the gallery structure but removing the hooks.
When you include both the mapper and placemapper there's some collision
happening with google maps instances. I'm sure it could be resolved with a
little digging, but considering I'm doing this all for a presentation _tonight_
I dont' think that digging is going to happen anytime soon ;)